Update message-form.jsx
I'll change the way that the page send messages. I want both people to connect to another peer that will receive the message that anyone sent, will disconnect from the person who sent it and then connect with the one who has to receive it and send it to them. In this way, it will be easier to make groups